Some refactoring, animation and path finding
I made it so the current cell has a different material then the rest of the cells and also if a cell is visited its material changes as well as if it has been backtracked through it. I also implemented a path finding algorithm that also creates a entrance and exit cells. In addition I refactored some of the code, adding some helper functions and regions in the maze generator class as it is getting pretty long.