Refactoring code
I refactored almost all classes. I made it more modular and reusable. I made separate classes for each algorithm as well as an interface for the algorithms. I also made an interface for a maze grid, so that it is possible to create a different type of grid that is not a hex one. I also made a base cell class from which the hex cell inherits. In addition I added some comments and summaries for each function.